requirements.txt
apache-airflow==2.10.5
pandas==2.3.3
plotly==6.3.1
python_dateutil==2.9.0.post0
streamlit==1.50.0
streamlit_autorefresh==1.0.1
ta==0.11.0
tenacity==9.1.2
vnstock==3.2.6
pyrate-limiter==3.9.0